2023.11.24

+++ scripts/install_pve.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # from
 # https://github.com/spiritLHLS/pve
-# 2023.11.22
+# 2023.11.24
 
 ########## 预设部分输出和部分中间变量
 
@@ -436,6 +436,21 @@ is_private_ipv6() {
 
 check_ipv6() {
     IPV6=$(ip -6 addr show | grep global | awk '{print length, $2}' | sort -nr | head -n 1 | awk '{print $2}' | cut -d '/' -f1)
+    ipv6_list=$(ip -6 addr show | grep global | awk '{print length, $2}' | sort -nr | awk '{print $2}')
+    line_count=$(echo "$ipv6_list" | wc -l)
+    if [ "$line_count" -ge 2 ]; then
+        # 获取最后一行的内容
+        last_ipv6=$(echo "$ipv6_list" | tail -n 1)
+        # 切分最后一个:之前的内容
+        last_ipv6_prefix="${last_ipv6%:*}:"
+        # 与${ipv6_gateway}比较是否相同
+        if [ "${last_ipv6_prefix}" = "${ipv6_gateway}" ]; then
+            echo $last_ipv6 >/usr/local/bin/pve_last_ipv6
+        fi
+        _green "The local machine is bound to more than one IPV6 address"
+        _green "本机绑定了不止一个IPV6地址"
+    fi
+
     if is_private_ipv6 "$IPV6"; then # 由于是内网IPV6地址，需要通过API获取外网地址
         IPV6=""
         API_NET=("ipv6.ip.sb" "https://ipget.net" "ipv6.ping0.cc" "https://api.my-ip.io/ip" "https://ipv6.icanhazip.com")
@@ -860,9 +875,6 @@ if [ ! -f /etc/systemd/network/10-persistent-net.link ]; then
 fi
 
 # 检测IPV6相关的信息
-if [ ! -f /usr/local/bin/pve_check_ipv6 ] || [ ! -s /usr/local/bin/pve_check_ipv6 ] || [ "$(sed -e '/^[[:space:]]*$/d' /usr/local/bin/pve_check_ipv6)" = "" ]; then
-    check_ipv6
-fi
 if [ ! -f /usr/local/bin/pve_ipv6_gateway ] || [ ! -s /usr/local/bin/pve_ipv6_gateway ] || [ "$(sed -e '/^[[:space:]]*$/d' /usr/local/bin/pve_ipv6_gateway)" = "" ]; then
     ipv6_gateway=$(ip -6 route show | awk '/default via/{print $3}' | head -n1)
     # output=$(ip -6 route show | awk '/default via/{print $3}')
@@ -880,6 +892,10 @@ if [ ! -f /usr/local/bin/pve_ipv6_gateway ] || [ ! -s /usr/local/bin/pve_ipv6_ga
     # fi
     echo "$ipv6_gateway" >/usr/local/bin/pve_ipv6_gateway
 fi
+ipv6_gateway=$(cat /usr/local/bin/pve_ipv6_gateway)
+if [ ! -f /usr/local/bin/pve_check_ipv6 ] || [ ! -s /usr/local/bin/pve_check_ipv6 ] || [ "$(sed -e '/^[[:space:]]*$/d' /usr/local/bin/pve_check_ipv6)" = "" ]; then
+    check_ipv6
+fi
 if [ ! -f /usr/local/bin/pve_fe80_address ] || [ ! -s /usr/local/bin/pve_fe80_address ] || [ "$(sed -e '/^[[:space:]]*$/d' /usr/local/bin/pve_fe80_address)" = "" ]; then
     fe80_address=$(ip -6 addr show dev $interface | awk '/inet6 fe80/ {print $2}')
     echo "$fe80_address" >/usr/local/bin/pve_fe80_address
@@ -891,7 +907,6 @@ else
     ipv6_gateway_fe80="N"
 fi
 ipv6_address=$(cat /usr/local/bin/pve_check_ipv6)
-ipv6_gateway=$(cat /usr/local/bin/pve_ipv6_gateway)
 fe80_address=$(cat /usr/local/bin/pve_fe80_address)
 if [ ! -f /usr/local/bin/pve_ipv6_prefixlen ] || [ ! -s /usr/local/bin/pve_ipv6_prefixlen ] || [ "$(sed -e '/^[[:space:]]*$/d' /usr/local/bin/pve_ipv6_prefixlen)" = "" ]; then
     ipv6_prefixlen=""
@@ -1294,7 +1309,10 @@ iface vmbr0 inet6 auto
     bridge_ports $interface
 EOF
     else
-        cat <<EOF | sudo tee -a /etc/network/interfaces
+        # 与${ipv6_gateway}比较是否相同
+        if [ -f /usr/local/bin/pve_last_ipv6 ]; then
+            last_ipv6=$(cat /usr/local/bin/pve_last_ipv6)
+            cat <<EOF | sudo tee -a /etc/network/interfaces
 auto vmbr0
 iface vmbr0 inet static
     address $ipv4_address
@@ -1304,9 +1322,27 @@ iface vmbr0 inet static
     bridge_fd 0
 
 iface vmbr0 inet6 static
-        address ${ipv6_address_without_last_segment}1/128
-        gateway ${ipv6_gateway}
+    address ${last_ipv6}
+    gateway ${ipv6_gateway}
+
+iface vmbr0 inet6 static
+    address ${ipv6_address_without_last_segment}1/128
 EOF
+        else:
+            cat <<EOF | sudo tee -a /etc/network/interfaces
+auto vmbr0
+iface vmbr0 inet static
+    address $ipv4_address
+    gateway $ipv4_gateway
+    bridge_ports $interface
+    bridge_stp off
+    bridge_fd 0
+
+iface vmbr0 inet6 static
+    address ${ipv6_address_without_last_segment}1/128
+    gateway ${ipv6_gateway}
+EOF
+        fi
     fi
 fi
 chattr +i /etc/network/interfaces
